# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for crypto-knight trading (tracks, entries, balance, orders, klines).",
        "title": "Crypto Knight API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tracks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "List tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From (YYYY-MM-DD HH:MI:SS)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To (YYYY-MM-DD HH:MI:SS)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Symbol",
                        "name": "symbol",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_track.Track"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Create track",
                "parameters": [
                    {
                        "description": "Track",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_track.Track"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Create tracks in bulk",
                "parameters": [
                    {
                        "description": "Tracks",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks-history"
                ],
                "summary": "List tracks history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From (YYYY-MM-DD HH:MI:SS)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To (YYYY-MM-DD HH:MI:SS)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Symbol",
                        "name": "symbol",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_track.Track"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks/history/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tracks-history"
                ],
                "summary": "Create tracks history in bulk",
                "parameters": [
                    {
                        "description": "Tracks",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks/last": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get last tracks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_track.Track"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "github_com_antongoncharik_crypto-knight-api_internal_entity_track.Track": {
            "type": "object",
            "required": [
                "interval",
                "symbol"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "high_price": {
                    "type": "number"
                },
                "high_prices": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "interval": {
                    "type": "string"
                },
                "low_price": {
                    "type": "number"
                },
                "low_prices": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "middle_price_high": {
                    "type": "number"
                },
                "middle_price_low": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
    }
}